pull publishtheme utils functions into own file
